carprice/requirements.txt
firebase_admin==6.5.0
numpy==2.0.0
pandas==2.2.2
Pillow==10.4.0
Requests==2.32.3
streamlit==1.36.0
streamlit_option_menu==0.3.13
scikit-learn
